+++ src/Feliz.CompilerPlugins/CHANGELOG.md
@@ -16,6 +16,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 3.0.0-rc.4 - 2025-11-18
+
+### 🗑️ Deprecated
+
+- Remove transformation of single input record types for ReactComponent #603 (by @Freymaurer)
+
+### 🐛 Fixed
+
+- Fix equality issue for single input record types for ReactComponent #603  (by @Freymaurer)
+
 ## 3.0.0-rc.3 - 2025-11-03
 
 ### 🐛 Fixed